[add] 1khz interrupt using GPTimer
