+++ dependencies.lock
@@ -12,7 +12,7 @@ dependencies:
   idf:
     source:
       type: idf
-    version: 5.3.2
+    version: 5.4.0
 direct_dependencies:
 - espressif/iqmath
 - idf
